ispravljena greška u algoritmu za čišćenje grafa

+++ scripts/clean_graph.py
@@ -1,7 +1,9 @@
-def clean_graph(G):
+def clean_graph(G, src, end):
     delete_one = []
     for node in G:
-        if len(G[node]) is 1:
+        if len(G[node]) is 1 and\
+                node is not src and\
+                node is not end:
             delete_one.append(node)
     G.remove_nodes_from(delete_one)
